clean up custom apps
