--- src/custom-apps/queries.ts
@@ -1,9 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const webhooksDetails = gql`
-  query WebhookDetails($id: ID!) {
-    webhook(id: $id) {
-      ...WebhookDetails
-    }
-  }
-`;
