+++ src/extensions/index.tsx
@@ -1,9 +1,9 @@
 import SectionRoute from "@dashboard/auth/components/SectionRoute";
 import { Route } from "@dashboard/components/Router";
 import { WindowTitle } from "@dashboard/components/WindowTitle";
-import { CustomAppDetailsUrlQueryParams } from "@dashboard/custom-apps/urls";
 import {
   AppDetailsUrlQueryParams,
+  CustomAppDetailsUrlQueryParams,
   ExtensionInstallQueryParams,
   ExtensionsPaths,
 } from "@dashboard/extensions/urls";
