+++ src/extensions/queries.ts
@@ -182,3 +182,11 @@ export const appWebhookDeliveries = gql`
     }
   }
 `;
+
+export const webhooksDetails = gql`
+  query WebhookDetails($id: ID!) {
+    webhook(id: $id) {
+      ...WebhookDetails
+    }
+  }
+`;
