+++ src/extensions/urls.ts
@@ -235,3 +235,12 @@ export const ExtensionsUrls = {
       })}`,
     ),
 };
+
+type CustomAppDetailsUrlDialog =
+  | "create-token"
+  | "remove-webhook"
+  | "remove-token"
+  | "app-activate"
+  | "app-deactivate";
+
+export type CustomAppDetailsUrlQueryParams = Dialog<CustomAppDetailsUrlDialog> & SingleAction;
