+++ src/extensions/utils.tsx
@@ -1,5 +1,10 @@
 import { gql } from "@apollo/client";
-import { AppErrorCode, AppErrorFragment } from "@dashboard/graphql";
+import {
+  AppErrorCode,
+  AppErrorFragment,
+  WebhookEventTypeAsyncEnum,
+  WebhookFragment,
+} from "@dashboard/graphql";
 import errorTracker from "@dashboard/services/errorTracking";
 import { IntlShape } from "react-intl";
 
@@ -97,7 +102,7 @@ export function getCustomAppErrorMessage(
   return undefined;
 }
 
-interface IntrospectionNode {
+export interface IntrospectionNode {
   name: string;
   interfaces: Array<{
     name: string;
@@ -124,3 +129,17 @@ const isEvent = ({ name }: { name: string }) => name === "Event";
 
 export const buildEventsMap = (elements: IntrospectionNode[]) =>
   elements.filter(({ interfaces }) => (interfaces || []).some(isEvent));
+
+export function isUnnamed(webhook: WebhookFragment | undefined): boolean {
+  return !webhook?.name;
+}
+
+export const filterSelectedAsyncEvents = (asyncEvents: WebhookEventTypeAsyncEnum[]) => {
+  const anyEvent = asyncEvents.find(event => event === WebhookEventTypeAsyncEnum.ANY_EVENTS);
+
+  if (anyEvent) {
+    return [anyEvent];
+  }
+
+  return asyncEvents;
+};
