+++ src/extensions/views/EditCustomExtension/components/WebhooksList/WebhooksList.tsx
@@ -3,8 +3,8 @@ import ResponsiveTable from "@dashboard/components/ResponsiveTable";
 import TableButtonWrapper from "@dashboard/components/TableButtonWrapper";
 import TableCellHeader from "@dashboard/components/TableCellHeader";
 import TableRowLink from "@dashboard/components/TableRowLink";
-import { isUnnamed } from "@dashboard/custom-apps/utils";
 import { ExtensionsUrls } from "@dashboard/extensions/urls";
+import { isUnnamed } from "@dashboard/extensions/utils";
 import { WebhookFragment } from "@dashboard/graphql";
 import useNavigator from "@dashboard/hooks/useNavigator";
 import { commonMessages, commonStatusMessages, sectionNames } from "@dashboard/intl";
